requirements.txt
Flask==2.3.2
keras==2.13.1
numpy==1.24.3
Pillow==10.0.0
Pillow==10.0.0
Pyrebase4==4.7.1
